Enable QUICKLZ safe mode
If a page is corrupted in the database (e.g., due to catastrophic write failure), it is possible that the compressed data is corrupted, which can cause quicklz to do bad things if safe mode is not enabled. This change will force quicklz to be careful with decompression, at the expense of performance.